Add Astronomy Picture of the Day from NASA.
Astronomy Picture of the Day (APOD) is one of the most popular websites
at NASA.[1] In fact, this website is one of the most popular websites
across all federal agencies.

This endpoint structures the APOD imagery and associated metadata
so that it can be repurposed for other applications.

In addition, if the concept_tags parameter is set to True, then keywords
derived from the image explanation are returned. These keywords could be
used as auto-generated hashtags for Twitter or Instagram feeds; but
generally help with discoverability of relevant imagery.

[1] https://api.nasa.gov/api.html#apod